pyproject.toml
[project]
name = "project-mining-laser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.49.0",
    "asyncio>=3.4.3",
    "beautifulsoup4>=4.13.3",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "langchain[openai]>=0.3.22",
    "langgraph-bigtool>=0.0.2",
    "logfire>=3.12.0",
    "marko>=2.1.2",
    "mcp[cli]>=1.6.0",
    "openai>=1.70.0",
    "psycopg2>=2.9.10",
    "pydantic-ai-slim[mcp]>=0.0.48",
]
